+++ app/src/main/java/ch/epfl/sdp/peakar/user/services/RemoteAuthAccount.java
@@ -55,8 +55,6 @@ public class RemoteAuthAccount extends AuthAccount implements RemoteResource {
         return remoteOutcome;
     }
 
-    /* GETTERS */
-
     @Override
     public void init() {
         if(getUsername().equals(USERNAME_BEFORE_REGISTRATION)) {
